requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic

# HTTP client
requests

# Configuration
python-dotenv

# Logging
termcolor
